Reduce default exposed ports to host to only DNS
By default only expose the minimum set of ports necessary to use the
services. At this point it is only the DNS 53 UDP port. The rest
are supposed to only be exposed to the host and other containers.

Even if we only wanted to expose this port for the host, this has
to be openned externally because Docker is currently built that way
(TCP is auto-exposed to host, UDP is not).

Doubt here: I'm assuming that exposing something to the host exposes
it to the outside world, but I would have to test this.