Add lightcontrol example
This is a Java example which uses Maven and Jersey to deploy two
standalone systems which communicate with each other and with the core
services.

Please note that this commit does not work without a follow-up commit
(the one right after this) which modifies the core services to be more
friendly towards external containers.