Add the base for a COAP core-services proxy
Currently only has ServiceDiscovery implemented (through
SimpleServiceDiscovery) but it should be relatively easy to add the
rest of the services, especially if some refactoring is done.

On the refactoring: only a few classes should be needed and not a whole
load a class per resource. Instead create a class for forwarding calls
and just use configurations and defaults to determine the mappings.
Much more maintainable.